hw-6 optimization
