+++ nvleonovich_homework/nvleonovich_homework/Friends/FriendsPhotoCollectionViewController.swift
@@ -1,5 +1,4 @@
 import UIKit
-//import SDWebImage
 import RealmSwift
