+++ nvleonovich_homework/nvleonovich_homework/News/NewsViewController.swift
@@ -21,23 +21,11 @@ class NewsViewController: UITableViewController {
     }
     
     private func requestData() {
-        DispatchQueue.main.async {
-            Requests.go.getNews { [weak self] news, sourceDetails in
-                self?.myNews = news
-                self?.sourceDetails = sourceDetails
-                self?.tableView.reloadData()
-            }
+        Requests.go.getNews { [weak self] news, sourceDetails in
+            self?.myNews = news
+            self?.sourceDetails = sourceDetails
+            self?.tableView.reloadData()
         }
-    }
-    
-    private func convertDate(_ timeStampDate: Int) -> String {
-        let date = Date(timeIntervalSince1970: TimeInterval(timeStampDate))
-        let dateFormatter = DateFormatter()
-        dateFormatter.timeStyle = DateFormatter.Style.medium //Set time style
-        dateFormatter.dateStyle = DateFormatter.Style.medium //Set date style
-        dateFormatter.timeZone = .current
-        let localDate = dateFormatter.string(from: date)
-        return localDate
     }
     
     private func getSourceName(_ index: Int) -> String {
@@ -73,12 +61,13 @@ class NewsViewController: UITableViewController {
         let cell = tableView.dequeueReusableCell(withIdentifier: "NewsCard", for: indexPath) as! NewsCard
         cell.ownerName.text = getSourceName(indexPath.row)
         let avatar = getSourceAvatar(indexPath.row)
-        cell.postOwnerAvatar.sd_setImage(with: URL(string: avatar), placeholderImage: UIImage(named: "placeholder-1-300x200.png"))
-//        cell.postPhoto.sd_setImage(with: URL(string: newsCard.photo ?? "placeholder-1-300x200.png"), placeholderImage: UIImage(named: "placeholder-1-300x200.png"))
+        
+        cell.postOwnerAvatar.image = imageService?.photo(atIndexpath: indexPath, byUrl: avatar )
         cell.postPhoto.image = imageService?.photo(atIndexpath: indexPath, byUrl: newsCard.photo ?? "placeholder-1-300x200.png")
 //        cell.postPhoto.image = myNews[indexPath.row].photo
     
-        cell.publishDate.text = convertDate(newsCard.publishDate)
+//        cell.publishDate.text = convertDate(newsCard.publishDate)
+        cell.publishDate.text = DateConverter.get.convertDate(newsCard.publishDate)
         cell.postDescription.text = newsCard.text
         cell.viewsCount.text = "\(newsCard.viewsCount)"
         cell.likesCount.text = "\(newsCard.likesCount)"
